requirements.txt
Flask
requests
python-dotenv
pydub
elevenlabs
streamlit